refactor: terrain shaping moves out of CityGenerator (#11)
gen/Terrain holds everything that reads or rewrites vanilla terrain
before the city goes in: clearing a range, flattening and blending at a
city's edge, walking a column up or down to the first solid or empty
block, and the height offsets that decide where a city level sits.

isEmpty stays behind. It is a predicate on a block state with 107 callers
across the mod - vocabulary rather than terrain shaping, and moving it
would be 107 edits for no structural gain.

Pure code motion, verified token-for-token: 901 and 682 tokens on both
sides of the two ranges.

A fifth blind-replacement corruption caught on the way, this one in a
declaration rather than a body: rewriting `provider` to `feature.provider`
renamed the parameter of isWaterBiome(PlanningContext provider, ...) and
produced `PlanningContext feature.provider`. That failed to compile, but
the same rewrite one character different would not have.

CityGenerator 1030 -> 731 lines, from 2425 at the start of this
milestone. What is left is chunk orchestration - generate,
doNormalChunk, doCityChunk, generateBuilding - plus the vocabulary the
gen package speaks.

All seven digest suites unchanged.
